+++ prj/inc/generator.hh
@@ -17,6 +17,6 @@
 * Funkcja generuje naturalne  liczby losowe z przedziału 0-100, ktore nastepnie sa zapisywane do pliku random_data.dat
 * \param[in] data_amount - ilosc liczb wynikowych ktore chcemy uzyskac
 */
-bool data_generator (int data_amount);
+bool data_generator (unsigned long int data_amount);
   
 #endif
